Program that finds and sums of even fibonacci numbers less than 4000000

+++ 0x02-functions_nested_loops/103-fibonacci.c
@@ -7,7 +7,7 @@
  * @sum: sum of even numbers
  * Return: 0 if successful
  */
-void phi(unsigned long x, unsigned long y, unsigned long long sum);
+void phi(unsigned long x, unsigned long y, unsigned long sum);
 int main(void)
 {
 	phi(2, 1, 0);
@@ -20,11 +20,11 @@ int main(void)
  * @y: one step behind x
  * @sum: sum of even numbers
  */
-void phi(unsigned long x, unsigned long y, unsigned long long sum)
+void phi(unsigned long x, unsigned long y, unsigned long sum)
 {
 	if (x < 4000000)
 	{
-		unsigned long long z = x;
+		unsigned long z = x;
 
 		if (x % 2 == 0)
 			sum += x;
@@ -34,7 +34,7 @@ void phi(unsigned long x, unsigned long y, unsigned long long sum)
 	}
 	else
 	{
-		printf("%llu", sum);
+		printf("%lu", sum);
 		printf("\n");
 	}
 }
